requirements.txt
aiofiles==22.1.0
aiohttp==3.7.4.post0
aiohttp-socks==0.7.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==3.0.1
atomicwrites==1.4.1
attrs==21.4.0
autocommand==2.2.1
btrfsutil==6.0
cachetools==5.2.0
cairocffi==1.4.0
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.0.0
cryptography==38.0.1
Cython==0.29.32
discord.py==2.0.1
frozenlist==1.3.1
future==0.18.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
inflect==6.0.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.10.0
Jinja2==3.1.2
jsonschema==4.16.0
Logbook==1.5.3
MarkupSafe==2.1.1
matrix==0.3.0
matrix-nio==0.20.0
more-itertools==9.0.0
multidict==6.0.2
numpy==1.23.4
ordered-set==4.1.0
packaging==21.3
peewee==3.15.3
ply==3.11
psutil==5.9.3
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
Pygments==2.13.0
PyGObject==3.42.2
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
pyrsistent==0.19.0
python-olm==3.2.12
python-socks==2.0.3
PyYAML==6.0
qtile==0.22.1
qutebrowser==2.5.2
ranger-fm==1.9.3
Reflector==2021.11.20.2.41.3
six==1.16.0
speedtest-cli==2.1.3
tomli==2.0.1
trash-cli==0.22.8.27
trove-classifiers==2022.10.19
typing_extensions==4.2.0
unpaddedbase64 @ file:///build/python-unpaddedbase64/src/python-unpaddedbase64-2.1.0/dist/unpaddedbase64-2.1.0-py3-none-any.whl
validate-pyproject==0.10.1
webcolors==1.12
xcffib==0.12.1
yarl==1.7.2
youtube-dl==2021.12.17
